addition of files: adding files for the project
This is the start of my portfolio website. I have added the basic
structure of the website  and the basic components and have used
tailwind css for the styling of the website, prisma for the database,
eslint for the linting of the code, typescript for the type checking of
the code and last but not least I'm using  nextjs for the server side
rendering of the website.

Signed-off-by: Ashutosh Pandey <ashutosh.pandeyhlr007@gmail.com>